+++ pybamm/models/full_battery_models/lead_acid/higher_order.py
@@ -49,6 +49,7 @@ class BaseHigherOrderModel(BaseModel):
         self.set_full_interface_submodel()
         self.set_full_convection_submodel()
         self.set_full_porosity_submodel()
+        self.set_active_material_submodel()
         self.set_tortuosity_submodels()
         self.set_thermal_submodel()
         self.set_current_collector_submodel()
@@ -134,12 +135,12 @@ class BaseHigherOrderModel(BaseModel):
         )
 
     def set_negative_electrode_submodel(self):
-        self.submodels["negative electrode"] = pybamm.electrode.ohm.Composite(
+        self.submodels["negative electrode potential"] = pybamm.electrode.ohm.Composite(
             self.param, "Negative"
         )
 
     def set_positive_electrode_submodel(self):
-        self.submodels["positive electrode"] = pybamm.electrode.ohm.Composite(
+        self.submodels["positive electrode potential"] = pybamm.electrode.ohm.Composite(
             self.param, "Positive"
         )
 
@@ -152,12 +153,16 @@ class BaseHigherOrderModel(BaseModel):
         self.submodels["negative interface"] = pybamm.interface.FirstOrderKinetics(
             self.param,
             "Negative",
-            pybamm.interface.ButlerVolmer(self.param, "Negative", "lead-acid main"),
+            pybamm.interface.ButlerVolmer(
+                self.param, "Negative", "lead-acid main", self.options
+            ),
         )
         self.submodels["positive interface"] = pybamm.interface.FirstOrderKinetics(
             self.param,
             "Positive",
-            pybamm.interface.ButlerVolmer(self.param, "Positive", "lead-acid main"),
+            pybamm.interface.ButlerVolmer(
+                self.param, "Positive", "lead-acid main", self.options
+            ),
         )
 
         # Oxygen
@@ -168,7 +173,7 @@ class BaseHigherOrderModel(BaseModel):
                 self.param,
                 "Positive",
                 pybamm.interface.ForwardTafel(
-                    self.param, "Positive", "lead-acid oxygen"
+                    self.param, "Positive", "lead-acid oxygen", self.options
                 ),
             )
             self.submodels[
@@ -182,7 +187,7 @@ class BaseHigherOrderModel(BaseModel):
         Update convection submodel, now that we have the spatially heterogeneous
         interfacial current densities
         """
-        if self.options["convection"] is not False:
+        if self.options["convection"] != "none":
             self.submodels[
                 "through-cell convection"
             ] = pybamm.convection.through_cell.Explicit(self.param)