+++ pybamm/models/submodels/interface/sei/reaction_limited.py
@@ -27,7 +27,7 @@ class ReactionLimited(BaseModel):
         L_outer = pybamm.standard_variables.L_outer
 
         variables = self._get_standard_thickness_variables(L_inner, L_outer)
-        variables.update(self._get_standard_concentraion_variables(variables))
+        variables.update(self._get_standard_concentration_variables(variables))
 
         return variables