+++ tests/unit/test_models/test_full_battery_models/test_base_battery_model.py
@@ -161,11 +161,32 @@ class TestBaseBatteryModel(unittest.TestCase):
             pybamm.BaseBatteryModel({"sei film resistance": "bad sei film resistance"})
         with self.assertRaisesRegex(pybamm.OptionError, "sei porosity change"):
             pybamm.BaseBatteryModel({"sei porosity change": "bad sei porosity change"})
-        # variable defaults
+        with self.assertRaisesRegex(
+            pybamm.OptionError, "sei porosity change must now be given in string format"
+        ):
+            pybamm.BaseBatteryModel({"sei porosity change": True})
+        # changing defaults based on other options
         model = pybamm.BaseBatteryModel()
-        self.assertEqual(model.options["sei film resistance"], None)
+        self.assertEqual(model.options["sei film resistance"], "none")
         model = pybamm.BaseBatteryModel({"sei": "constant"})
         self.assertEqual(model.options["sei film resistance"], "distributed")
+        self.assertEqual(
+            model.options["total interfacial current density as a state"], "true"
+        )
+        with self.assertRaisesRegex(pybamm.OptionError, "must be 'true'"):
+            model = pybamm.BaseBatteryModel(
+                {
+                    "sei film resistance": "distributed",
+                    "total interfacial current density as a state": "false",
+                }
+            )
+
+        # loss of active material model
+        with self.assertRaisesRegex(pybamm.OptionError, "loss of active material"):
+            model = pybamm.BaseBatteryModel(
+                {"loss of active material": "bad LAM model"}
+            )
+
         # crack model
         with self.assertRaisesRegex(pybamm.OptionError, "particle cracking"):
             pybamm.BaseBatteryModel({"particle cracking": "bad particle cracking"})