+++ tests/unit/test_models/test_submodels/test_interface/test_kinetics/test_tafel.py
@@ -26,8 +26,17 @@ class TestTafel(unittest.TestCase):
             "Negative electrolyte concentration": a_n,
             "Negative particle surface concentration": a_n,
             "Negative electrode temperature": a_n,
+            "Negative electrode surface area to volume ratio": a_n,
         }
-        submodel = pybamm.interface.ForwardTafel(param, "Negative", "lithium-ion main")
+        submodel = pybamm.interface.ForwardTafel(
+            param,
+            "Negative",
+            "lithium-ion main",
+            {
+                "sei film resistance": "none",
+                "total interfacial current density as a state": "false",
+            },
+        )
         std_tests = tests.StandardSubModelTests(submodel, variables)
 
         std_tests.test_all()
@@ -43,6 +52,8 @@ class TestTafel(unittest.TestCase):
             "Negative electrode interfacial current density": a_n,
             "Negative electrode exchange current density": a_n,
             "Positive electrode temperature": a_p,
+            "Negative electrode surface area to volume ratio": a_n,
+            "Positive electrode surface area to volume ratio": a_p,
             "X-averaged negative electrode interfacial current density": a,
             "X-averaged positive electrode interfacial current density": a,
             "Sum of electrolyte reaction source terms": 0,
@@ -58,7 +69,15 @@ class TestTafel(unittest.TestCase):
             "Sum of x-averaged negative electrode interfacial current densities": 0,
             "Sum of x-averaged positive electrode interfacial current densities": 0,
         }
-        submodel = pybamm.interface.BackwardTafel(param, "Positive", "lithium-ion main")
+        submodel = pybamm.interface.BackwardTafel(
+            param,
+            "Positive",
+            "lithium-ion main",
+            {
+                "sei film resistance": "none",
+                "total interfacial current density as a state": "false",
+            },
+        )
         std_tests = tests.StandardSubModelTests(submodel, variables)
         std_tests.test_all()