+++ tests/unit/test_models/test_submodels/test_particle/test_fickian_single_particle.py
@@ -15,6 +15,8 @@ class TestSingleParticle(unittest.TestCase):
         variables = {
             "X-averaged negative electrode interfacial current density": a,
             "X-averaged negative electrode temperature": a,
+            "Negative electrode active material volume fraction": a,
+            "Negative electrode surface area to volume ratio": a,
         }
 
         submodel = pybamm.particle.FickianSingleParticle(param, "Negative")
@@ -24,6 +26,8 @@ class TestSingleParticle(unittest.TestCase):
         variables = {
             "X-averaged positive electrode interfacial current density": a,
             "X-averaged positive electrode temperature": a,
+            "Positive electrode active material volume fraction": a,
+            "Positive electrode surface area to volume ratio": a,
         }
         submodel = pybamm.particle.FickianSingleParticle(param, "Positive")
         std_tests = tests.StandardSubModelTests(submodel, variables)